safety-app/voice-agent/requirements.txt
livekit-agents[silero,turn-detector]>=1.2.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv>=1.0.0
